Add live preview
This gets compiled out in production (since @ifsimicoded rightly pointed out this could be an easy target for attackers), but will allow joplin to pass over URL params to set the correct data.

I've been testing this locally and I don't seem to run into any length limitations. If that happens we'll need to compress the data, but I don't want to do that before it's necessary.