Fix active state in Navmenu
Active state styling wasn't being applied in the Navmenu because of language code prefixed in the paths.

Doing this made me think that the NavLink component provided by react-router-dom is a better solution than this. ex: #102 🙃

We should use existing router functionality once we choose a router for server-side/static generation.