Update to latest react-static
This solves a build problem I was running into, though I'm not entirely sure why. It seems like it's a `TileGroup` or `TileGroupSet` issue, but I'm not ready to track it down.

The old version would fail the build but the new version shows the following warning

```
Check the top-level render call using <div>. See https://fb.me/react-warning-keys for more information.
    in TileGroup
    in TileGroupSet
    in div
    in div
    in Theme
...
```